Add files via upload
Problem Statement:
We have Collected data From various research of what kind of books people like or dislike. On the basis of the data collected we have to build a recommendation model which would recommend the books to the user based on their previous purchased or similar liking books. For that Build a recommender system by using cosine similarities score.